@@ -4,10 +4,10 @@
   "private": true,
   "dependencies": {
     "classnames": "^2.2.6",
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
-    "react-jss": "^8.6.1",
-    "react-scripts": "1.1.4"
+    "node-sass": "^4.13.0",
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0",
+    "react-scripts": "^3.2.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -18,5 +18,17 @@
   "devDependencies": {
     "eslint-config-airbnb": "^17.1.0",
     "prettier-eslint": "^8.8.2"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
