@@ -1,44 +1,36 @@
-import React, { Component } from "react";
-import injectSheet from "react-jss";
+import React, { useState } from "react";
 import { Text, Search, UserCard } from "../../ui";
-import { styles } from "./Users.style";
+import classes from "./Users.module.scss";
 
-class Users extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      users: props.users
-    };
-  }
-  onChange = e => {
+const Users = ({ users }) => {
+  const [filteredUsers, updateFilteredUsers] = useState(users);
+
+  const onChange = e => {
     const regex = new RegExp(e.target.value.toLowerCase());
-    const users = this.props.users.filter(({ name }) =>
+    const filteredUsers = users.filter(({ name }) =>
       regex.test(name.toLowerCase())
     );
-    this.setState({ users });
+    updateFilteredUsers(filteredUsers);
   };
-  render() {
-    const { classes } = this.props;
-    return (
-      <div className={classes.wrapper}>
-        <div className={classes.heading}>
-          <Search className={classes.search} onChange={this.onChange} />
-          <Text type="heading" className={classes.title}>
-            Users
-          </Text>
-        </div>
-        <div className={classes.content}>
-          <div className={classes.userCards}>
-            {this.state.users.map(user => (
-              <UserCard key={Math.random()} user={user} />
-            ))}
-          </div>
-          <div className={classes.userStats}>Stats</div>
-        </div>
+  return (
+    <div className={classes.wrapper}>
+      <div className={classes.heading}>
+        <Search className={classes.search} onChange={onChange} />
+        <Text type="heading" className={classes.title}>
+          Users
+        </Text>
       </div>
-    );
-  }
-}
+      <div className={classes.content}>
+        <div className={classes.userCards}>
+          {filteredUsers.map(user => (
+            <UserCard key={Math.random()} user={user} />
+          ))}
+        </div>
+        <div className={classes.userStats}>Stats</div>
+      </div>
+    </div>
+  );
+};
 
 Users.defaultProps = { users: [{ name: "Jon Rose" }, { name: "Yen Rose" }] };
-export default injectSheet(styles)(Users);
+export default Users;
