@@ -1,8 +1,4 @@
-import React, { Component } from 'react';
-import Users from './Users.component';
+import React from "react";
+import Users from "./Users.component";
 
-export default class UsersContainer extends Component {
-    render() {
-        return(<Users />)
-    }
-}
+export default _ => <Users />;
