@@ -1,10 +1,9 @@
-import React from 'react';
-import injectSheet from 'react-jss'
-import classNames from 'classnames';
-import { styles } from './Text.style';
+import React from "react";
+import classNames from "classnames";
+import classes from "./Text.module.scss";
 
-const text = ({ classes, type, children, className }) => (
-    <p className={classNames(classes.text, classes[type], className)}>{children}</p>
+export default ({ type, children, className }) => (
+  <p className={classNames(classes.text, classes[type], className)}>
+    {children}
+  </p>
 );
-
-export default injectSheet(styles)(text);
